+++ addon/button.js
@@ -15,7 +15,6 @@ if (document.querySelector("body.sfdcBody, body.ApexCSIPage, #auraLoadingBox") |
 }
 
 function initButton(sfHost, inInspector) {
-  addFlowScrollability();
   let rootEl = document.createElement("div");
   rootEl.id = "insext";
   let btn = document.createElement("div");
@@ -33,11 +32,19 @@ function initButton(sfHost, inInspector) {
     loadPopup();
   });
 
+  addFlowScrollability();
+
 
   function addFlowScrollability() {
     const currentUrl = window.location.href;
     // Check the current URL for the string "builder_platform_interaction"
     if (currentUrl.includes("builder_platform_interaction")) {
+      //add marging for the popup arrow to prevent overlap with standard close button in flow builder (Winter 24)
+      //temporary workaround, will be removed in next release when the popupArrow position will be updatable by users
+      const popupArrow = document.querySelector("#insext");
+      if (popupArrow){
+        popupArrow.style = "margin-top: 50px;";
+      }
       // Create a new checkbox element
       const headerFlow = document.querySelector("builder_platform_interaction-container-common");
       const overflowCheckbox = document.createElement("input");